Updated files of exceptions

+++ src/menu/menu1.controller.ts
@@ -1,15 +1,24 @@
 import { InjectModel } from '@nestjs/azure-database';
-import { Body, Controller, Delete, Put, Query } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  HttpException,
+  Put,
+  Query,
+  UseFilters,
+} from '@nestjs/common';
 import { Container } from '@azure/cosmos';
 import { MenuDto } from './menu.dto';
 import { menu } from './menu.entity';
-
+import { MenuExceptionFilter } from 'src/Exceptions/menu-exception.filter';
 
 @Controller('menu1')
+@UseFilters(MenuExceptionFilter) // Use your custom exception filter
 export class Menu1Controller {
   constructor(@InjectModel(menu) private readonly menuContainer: Container) {}
 
-  //Update an item in the database
+  // Update an item in the database
   @Put('update')
   async update(@Body() payload: MenuDto) {
     const itemupdate = new menu();
@@ -20,6 +29,7 @@ export class Menu1Controller {
     itemupdate.price = payload.price;
 
     const { resource } = await this.menuContainer.items.upsert(itemupdate);
+
     return {
       id: resource.id,
       outlet_name: resource.outlet_name,
@@ -29,14 +39,17 @@ export class Menu1Controller {
     };
   }
 
-  //Delete an item from the database
+  // Delete an item from the database
   @Delete('remove')
   async remove(
     @Query('id') id: string,
     @Query('partitionkey') partitionkey: string,
   ) {
+    if (parseInt(id) < 0 || parseInt(id) > 10 ) {
+      throw new HttpException('Enter ID greater than 0 less than 10', 400);
+    }
+    
     await this.menuContainer.item(id, partitionkey).delete();
     return 'deleted';
   }
-  
 }
